chore(kuma): move ResourceStatus to KumaResourceStatus
